Setup middleware to be ran standalone by express if not using webpack-dev-server; Separate api-products routes to standalone module; Install nodemon to make further API development easier;
