+++ package.json
@@ -22,6 +22,7 @@
     "test": "echo \"Tests not ready yet!\"",
     "prettier": "prettier --write src/ middleware/ database/ !database/data",
     "start": "webpack-dev-server",
+    "serve": "nodemon middleware/index.js -- --standalone=true",
     "prebuild": "npm run prettier",
     "build": "webpack"
   },
@@ -36,6 +37,7 @@
     "glob": "^7.1.6",
     "mini-css-extract-plugin": "^0.10.0",
     "node-sass": "^4.14.1",
+    "nodemon": "^2.0.4",
     "prettier": "^2.0.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
